4rd-project/chatbot_website/requirements.txt
absl-py==2.0.0
accelerate==0.28.0
ace_tools==0.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
alembic==1.13.3
altair==5.5.0
anaconda-client==1.11.1
anaconda-navigator==2.4.0
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
backports.weakref==1.0.post1
bayesian-optimization==1.5.1
bcrypt==4.2.0
benepar==0.2.0
bitsandbytes==0.42.0
blinker==1.8.2
blis==0.7.11
branca==0.6.0
brotlipy==0.7.0
build==1.2.2.post1
cachetools==5.3.1
catalogue==2.0.10
category-encoders==2.6.4
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
clyent==1.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
conda==23.3.1
conda-build==3.23.3
conda-repo-cli==1.0.41
conda-verify==3.4.2
confection==0.1.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.8
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.3.1
dask-expr==1.0.5
dataclasses==0.6
dataclasses-json==0.5.14
datasets==2.20.0
db-dtypes==1.2.0
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
Django==5.1.3
duckduckgo_search==6.3.6
durationpy==0.9
emoji==2.14.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
faiss-cpu==1.9.0
fastapi==0.115.4
fastjsonschema==2.20.0
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-ngrok==0.0.25
flatbuffers==24.3.25
folium==0.14.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.4.0
gdown==5.2.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
gluonnlp==0.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.18.0
google-api-python-client==2.151.0
google-auth==2.23.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.65.0
graphviz==0.8.4
groq==0.12.0
grpcio==1.67.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.12.1
haversine==2.8.1
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==7.1.0
importlib_resources==6.4.5
iopath==0.1.10
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jiter==0.7.0
joblib==1.2.0
JPype1==1.5.0
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kagglehub==0.3.4
kaleido==0.2.1
keras==3.7.0
kiwisolver==1.4.4
konlpy==0.6.0
Korpora==0.2.0
kubernetes==31.0.0
lab==8.2
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-google-genai==2.0.4
langchain-groq==0.2.1
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langcodes==3.3.0
langdetect==1.0.9
langgraph==0.2.46
langgraph-checkpoint==2.0.3
langgraph-sdk==0.1.35
langsmith==0.1.140
layoutparser==0.3.4
libclang==16.0.6
lightgbm==4.5.0
lightning-utilities==0.11.8
llvmlite==0.43.0
locket==1.0.0
Mako==1.3.5
Markdown==3.4.4
markdown-it-py==3.0.0
marshmallow==3.21.3
matplotlib==3.7.1
mdurl==0.1.2
mecab-python==1.0.0
mecab-python3==1.0.10
missingno==0.5.2
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mxnet==1.6.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.2
navigator-updater==0.3.0
nbformat==5.4.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.2.2
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnx==1.17.0
onnxruntime==1.20.0
openai==1.54.3
openapi-schema-pydantic==1.2.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt-einsum==3.3.0
optree==0.13.1
optuna==4.0.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandas-gbq==0.22.0
partd==1.4.1
patsy==0.5.6
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.20.0
pillow==11.0.0
pip-tools==7.4.1
pip-upgrade==0.0.6
platformdirs==4.3.6
plotly==5.14.1
plotly-resampler==0.10.0
ply==3.11
pmdarima==2.0.4
portalocker==2.10.1
posthog==3.7.0
preshed==3.0.9
primp==0.8.0
proto-plus==1.23.0
protobuf==4.25.5
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.9.2
pycaret==3.3.2
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydata-google-auth==1.8.2
pydeck==0.9.1
pyDeprecate==0.3.2
pydot==3.0.2
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.1
pyod==2.0.2
pyparsing==3.0.9
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5-sip==12.11.0
python-decouple==3.8
python-dotenv==1.0.1
python-iso639==2024.10.22
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytorch-lightning==2.4.0
pyzmq==26.1.1
qtconsole==5.5.2
QtPy==2.4.1
RapidFuzz==3.10.1
ratsnlp==1.0.53
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.2
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.5.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.3
sentence-transformers==3.3.0
sentencepiece==0.2.0
simplejson==3.19.3
six==1.16.0
sktime==0.26.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlparse==0.5.2
squarify==0.4.3
srsly==2.4.8
starlette==0.41.2
statsmodels==0.14.4
streamlit==1.40.2
sympy==1.13.1
tavily-python==0.5.0
tbats==1.1.3
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.1
tensorflow==2.18.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.13.0
termcolor==2.3.0
tf_keras==2.18.0
thinc==8.2.3
threadpoolctl==3.1.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
tomli==2.0.1
torch==2.5.1
torch-struct==0.5
torchmetrics==1.5.1
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
tsdownsample==0.1.3
txt2tags==3.9
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ucimlrepo==0.0.7
unstructured==0.16.5
unstructured-client==0.27.0
unstructured-inference==0.8.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
wasabi==1.1.2
watchfiles==0.24.0
weasel==0.3.4
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.0
Werkzeug==3.0.4
wget==3.2
widgetsnbextension==4.0.11
wordcloud==1.9.2
wrapt==1.15.0
wurlitzer==3.1.1
xgboost==2.1.1
xlrd==2.0.1
xxhash==3.4.1
yarl==1.9.4
yellowbrick==1.5
zipp==3.18.1
